+++ packages/director/src/index.ts
@@ -2,6 +2,7 @@ export * from "./project";
 export * from "./tools";
 export * from "./highlight";
 export * from "./edits";
+export * from "./tracks";
 export * from "./filler";
 export * from "./slideshow";
 export * from "./demo";